+++ application/src/App.tsx
@@ -8,9 +8,9 @@ import ScanningTab from "./scouter/scanner/ScanningTab";
 import GeneralTab from "./strategy/general-tab/GeneralTab";
 import TeamTab from "./strategy/team-tab/TeamTab";
 import ScouterPreMatch from "./scouter/tabs/ScouterPreMatch";
-import ScouterTeleoperated, {
+import ScouterPartialTeleoperated, {
   ScouterTelePick,
-} from "./scouter/tabs/ScouterTeleoperated";
+} from "./scouter/tabs/ScouterPartialTeleoperated";
 import ScouterPostMatch from "./scouter/tabs/ScouterPostMatch";
 import StrategyTeleoperated from "./strategy/team-tab/sections/StrategyTeleoperated";
 import StrategyAutonomous from "./strategy/team-tab/sections/StrategyAutonomous";
@@ -30,7 +30,7 @@ import TestLinear from "./strategy/team-tab/sections/TestLinear";
 import Leaderboard from "./components/Leaderboard";
 import TestLinear2 from "./strategy/team-tab/sections/TestLinear2";
 import ScouterStats from "./strategy/ScouterStats";
-import ScouterTeleoperatedFixed from "./scouter/tabs/ScouterTeleOperatedFixed";
+import ScouterTeleoperated from "./scouter/tabs/ScouterTeleOperated";
 
 function getHiddenImage(path: string) {
   return (
@@ -129,7 +129,7 @@ const App: React.FC = () => {
               <Route path="matches" element={<MatchList />} />
               <Route path="scouting" element={<ScouterTab />}>
                 <Route path="prematch" element={<ScouterPreMatch />} />
-                <Route path="teleoperated" element={<ScouterTeleoperatedFixed />}>
+                <Route path="teleoperated" element={<ScouterTeleoperated />}>
                   <Route path="pick" element={<ScouterTelePick />} />
                 </Route>
                 <Route path="autonomous" element={<ScouterAutonomous />}>
