+++ application/src/scouter/tabs/ScouterTeleoperated.tsx
@@ -1,13 +1,19 @@
 import React from "react";
 import ScouterInputs from "../ScouterInputs";
 import { Outlet } from "react-router-dom";
+import ScouterPartialTeleoperated from "./ScouterPartialTeleoperated";
+import { blue } from "@mui/material/colors";
 
 const ScouterTeleoperated: React.FC = () => {
-  return <Outlet />;
-};
-
-export const ScouterTelePick: React.FC = () => {
-  return ScouterInputs.teleReefPick.create();
+  return <>
+    <ScouterPartialTeleoperated />
+    <div className="defens" style={ {backgroundColor: "#263163"} }>
+        {ScouterInputs.create([
+            ScouterInputs.defense,
+            ScouterInputs.defensiveEvasion,
+          ])}
+    </div>
+  </>
 };
 
 export default ScouterTeleoperated;
